feat(bindings): expose completion handle to foreign language
