+++ data-plane/bindings/adapter/src/completion_handle.rs
@@ -149,3 +149,216 @@ impl CompletionHandle {
         }
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use std::sync::Arc;
+
+    use crate::SlimError;
+    use slim_session::SessionError as SlimSessionError;
+
+    /// Test CompletionHandle basic functionality
+    #[tokio::test]
+    async fn test_completion_handle_success() {
+        // Create a successful completion
+        let (tx, rx) = tokio::sync::oneshot::channel();
+        let completion = slim_session::CompletionHandle::from_oneshot_receiver(rx);
+        let ffi_handle = Arc::new(crate::CompletionHandle::from(completion));
+
+        // Send success
+        tx.send(Ok(())).unwrap();
+
+        // Wait should succeed
+        let result = ffi_handle.wait_async().await;
+        assert!(result.is_ok(), "Completion should succeed");
+    }
+
+    /// Test CompletionHandle failure propagation
+    #[tokio::test]
+    async fn test_completion_handle_failure() {
+        // Create a failed completion
+        let (tx, rx) = tokio::sync::oneshot::channel();
+        let completion = slim_session::CompletionHandle::from_oneshot_receiver(rx);
+        let ffi_handle = Arc::new(crate::CompletionHandle::from(completion));
+
+        // Send error
+        tx.send(Err(slim_session::SessionError::SlimMessageSendFailed))
+            .unwrap();
+
+        // Wait should fail with error
+        let result = ffi_handle.wait_async().await;
+        assert!(result.is_err_and(|e| matches!(e, SlimError::SessionError { .. })));
+    }
+
+    /// Test CompletionHandle can only be consumed once
+    #[tokio::test]
+    async fn test_completion_handle_single_consumption() {
+        let (tx, rx) = tokio::sync::oneshot::channel();
+        let completion = slim_session::CompletionHandle::from_oneshot_receiver(rx);
+        let ffi_handle = Arc::new(crate::CompletionHandle::from(completion));
+
+        tx.send(Ok(())).unwrap();
+
+        // First wait should succeed
+        let result1 = ffi_handle.clone().wait_async().await;
+        assert!(result1.is_ok(), "First wait should succeed");
+
+        // Second wait should fail (already consumed)
+        let result2 = ffi_handle.clone().wait_async().await;
+        assert!(result2.is_err(), "Second wait should fail");
+
+        match result2 {
+            Err(SlimError::InternalError { message }) => {
+                assert!(message.contains("already consumed"));
+            }
+            _ => panic!("Expected InternalError about consumption"),
+        }
+    }
+
+    /// Test CompletionHandle async version
+    #[tokio::test]
+    async fn test_completion_handle_async() {
+        let (tx, rx) = tokio::sync::oneshot::channel();
+        let completion = slim_session::CompletionHandle::from_oneshot_receiver(rx);
+        let ffi_handle = Arc::new(crate::CompletionHandle::from(completion));
+
+        // Send success in a separate task
+        tokio::spawn(async move {
+            tokio::time::sleep(std::time::Duration::from_millis(10)).await;
+            tx.send(Ok(())).unwrap();
+        });
+
+        // Wait for completion
+        let result = ffi_handle.wait_async().await;
+        assert!(result.is_ok(), "Async wait should succeed");
+    }
+
+    /// Test CompletionHandle with dropped sender
+    #[tokio::test]
+    async fn test_completion_handle_sender_dropped() {
+        let (_tx, rx) = tokio::sync::oneshot::channel();
+        let completion = slim_session::CompletionHandle::from_oneshot_receiver(rx);
+        let ffi_handle = Arc::new(crate::CompletionHandle::from(completion));
+
+        // Drop the sender explicitly
+        drop(_tx);
+
+        // Give the spawned task time to process
+        tokio::time::sleep(std::time::Duration::from_millis(50)).await;
+
+        // Wait should fail with session error
+        let result = ffi_handle.wait_async().await;
+        assert!(
+            result.is_err(),
+            "Should fail when sender is dropped or already consumed"
+        );
+
+        // Either error type is valid depending on timing
+        match result {
+            Err(SlimError::InternalError { message }) => {
+                assert!(
+                    message.contains("sender dropped") || message.contains("already consumed"),
+                    "Error message should mention sender dropped or already consumed, got: {}",
+                    message
+                );
+            }
+            Err(SlimError::SessionError { message }) => {
+                // The background task may have consumed the receiver and got a channel closed error
+                assert!(
+                    message.contains("channel closed") || message.contains("receiving ack"),
+                    "Expected channel closed error, got: {}",
+                    message
+                );
+            }
+            Err(e) => panic!("Unexpected error type: {:?}", e),
+            Ok(_) => panic!("Expected error, got Ok"),
+        }
+    }
+
+    /// Test concurrent completion handle usage
+    #[tokio::test]
+    async fn test_completion_handle_concurrent() {
+        use std::sync::atomic::{AtomicU32, Ordering};
+
+        let success_count = Arc::new(AtomicU32::new(0));
+        let mut handles = vec![];
+
+        // Create multiple completion handles
+        for _ in 0..10 {
+            let (tx, rx) = tokio::sync::oneshot::channel();
+            let completion = slim_session::CompletionHandle::from_oneshot_receiver(rx);
+            let ffi_handle = Arc::new(crate::CompletionHandle::from(completion));
+
+            let count = Arc::clone(&success_count);
+            let handle = tokio::spawn(async move {
+                // Send success
+                tx.send(Ok(())).unwrap();
+
+                // Wait for completion
+                if ffi_handle.wait_async().await.is_ok() {
+                    count.fetch_add(1, Ordering::SeqCst);
+                }
+            });
+
+            handles.push(handle);
+        }
+
+        // Wait for all tasks
+        for handle in handles {
+            handle.await.unwrap();
+        }
+
+        // All should succeed
+        assert_eq!(success_count.load(Ordering::SeqCst), 10);
+    }
+
+    /// Test CompletionHandle with timeout
+    #[tokio::test]
+    async fn test_completion_handle_with_timeout() {
+        // Create a completion that will never complete (sender not sent)
+        // NOTE: We must hold onto `tx` so the channel stays open.
+        // If we use `_tx`, it gets dropped immediately and the receiver
+        // returns RecvError instead of timing out.
+        let (tx, rx) = tokio::sync::oneshot::channel::<Result<(), SlimSessionError>>();
+        let completion = slim_session::CompletionHandle::from_oneshot_receiver(rx);
+        let ffi_handle = Arc::new(crate::CompletionHandle::from(completion));
+
+        // Wait with a short timeout - should timeout
+        let result = ffi_handle
+            .wait_for_async(std::time::Duration::from_millis(50))
+            .await;
+        assert!(
+            result.is_err(),
+            "Should timeout when operation doesn't complete"
+        );
+
+        match result {
+            Err(SlimError::Timeout) => {} // Expected
+            Err(e) => panic!("Expected Timeout error, got: {:?}", e),
+            Ok(_) => panic!("Expected timeout, got Ok"),
+        }
+
+        // Explicitly drop tx after the test to avoid unused variable warning
+        drop(tx);
+    }
+
+    /// Test CompletionHandle with timeout that completes in time
+    #[tokio::test]
+    async fn test_completion_handle_timeout_success() {
+        let (tx, rx) = tokio::sync::oneshot::channel();
+        let completion = slim_session::CompletionHandle::from_oneshot_receiver(rx);
+        let ffi_handle = Arc::new(crate::CompletionHandle::from(completion));
+
+        // Send success immediately
+        tx.send(Ok(())).unwrap();
+
+        // Wait with a generous timeout - should succeed
+        let result = ffi_handle
+            .wait_for_async(std::time::Duration::from_millis(5000))
+            .await;
+        assert!(
+            result.is_ok(),
+            "Should succeed when operation completes before timeout"
+        );
+    }
+}
